新刪修查ok

+++ WebApplication5/Models/OrderDetails.cs
@@ -25,16 +25,16 @@ namespace WebApplication5.Models
         /// <summary>
         /// 單價
         /// </summary>
-        public int UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
         /// <summary>
         /// 數量
         /// </summary>
-        public decimal Qty { get; set; }
+        public int Qty { get; set; }
 
         /// <summary>
         /// 折扣
         /// </summary>
-        public int Discount { get; set; }
+        public decimal Discount { get; set; }
     }
 }
